Fix bug where nested transactions would retry
If two transactions on separate connections were nested, the inner transaction would retry because the transaction count on its connection did not include the outer transaction on the other connection. If the inner transaction retried and failed, the outer transaction would rescue the inner transaction's `TransactionIsolationConflict` exception, and retry its block, which includes the inner transaction again. So two nested transactions that each retry once could cause the inner transaction to execute as many as four times, with two inner transaction attempts for each attempt of the outer transaction.

Instead, we now track how deeply nested our `transaction_with_rety` calls are as secondary indication of whether we are in a nested transaction, since the outer will retry the inner transaction, even if it is on a separate connection and isn't included in `ActiveRecord::Base.connection.open_transactions`.